Renames software
